# openapi/gateway.v1.yaml
openapi: 3.0.3
info:
  title: ApptRemind Gateway API
  version: v1
servers:
  - url: http://localhost:8080
paths:
...
  /api/v1/business/profile:
    get:
      summary: Get business profile
      security:
        - bearerAuth: []
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BusinessProfile"
              examples:
                default:
                  value:
                    business_id: "9f5f9e1a-7f8d-4b9c-9f7b-1e8f0c1d2e3f"
                    name: "Demo Salon"
                    timezone: "UTC"
                    reminder_offsets_minutes: [1440, 60]
    put:
      summary: Update business profile
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/BusinessProfileUpdateRequest"
            examples:
              update:
                value:
                  name: "Demo Salon"
                  timezone: "UTC"
                  reminder_offsets_minutes: [60]
      responses:
        "204":
          description: No Content
  /api/v1/business/services:
    post:
      summary: Create service
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/BusinessServiceCreateRequest"
            examples:
              consult:
                value:
                  name: "Consultation"
                  duration_minutes: 30
                  price: 25
                  description: "Initial consult"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/IdResponse"
              examples:
                created:
                  value:
                    id: "c6b6b7e0-7c2a-4a07-8a9f-1b2c3d4e5f60"
    get:
      summary: List services
      security:
        - bearerAuth: []
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/BusinessService"
              examples:
                list:
                  value:
                    - id: "c6b6b7e0-7c2a-4a07-8a9f-1b2c3d4e5f60"
                      business_id: "9f5f9e1a-7f8d-4b9c-9f7b-1e8f0c1d2e3f"
                      name: "Consultation"
                      duration_minutes: 30
                      price: "25.00"
                      description: "Initial consult"
                      created_at: "2026-01-28T10:00:00Z"
  /api/v1/business/staff:
    post:
      summary: Add staff member
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/StaffCreateRequest"
            examples:
              staff:
                value:
                  name: "Alice"
                  is_active: true
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/IdResponse"
              examples:
                created:
                  value:
                    id: "2d7f53f6-0b5f-4d0d-8f49-7a9c3b3b9c2a"
    get:
      summary: List staff
      security:
        - bearerAuth: []
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Staff"
              examples:
                list:
                  value:
                    - id: "2d7f53f6-0b5f-4d0d-8f49-7a9c3b3b9c2a"
                      business_id: "9f5f9e1a-7f8d-4b9c-9f7b-1e8f0c1d2e3f"
                      name: "Alice"
                      is_active: true
  /api/v1/business/staff/working-hours:
    get:
      summary: List staff working hours (0=Sun..6=Sat)
      security:
        - bearerAuth: []
      parameters:
        - name: staff_id
          in: query
          required: true
          schema:
            type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/WorkingHours"
              examples:
                list:
                  value:
                    - staff_id: "2d7f53f6-0b5f-4d0d-8f49-7a9c3b3b9c2a"
                      weekday: 1
                      is_working: true
                      start_minute: 540
                      end_minute: 1020
    put:
      summary: Upsert a staff working-hours row
      security:
        - bearerAuth: []
      parameters:
        - name: staff_id
          in: query
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/WorkingHoursUpsertRequest"
            examples:
              sunday:
                value:
                  weekday: 0
                  is_working: true
                  start_minute: 600
                  end_minute: 840
      responses:
        "204":
          description: No Content
  /api/v1/business/staff/time-off:
    post:
      summary: Create staff time off (blackout)
      security:
        - bearerAuth: []
      parameters:
        - name: staff_id
          in: query
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TimeOffCreateRequest"
            examples:
              blackout:
                value:
                  start_time: "2026-02-01T11:00:00Z"
                  end_time: "2026-02-01T12:00:00Z"
                  reason: "Doctor appointment"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/IdResponse"
              examples:
                created:
                  value:
                    id: "1f6b7a2c-9a3b-4a1e-8b2c-4c5d6e7f8a90"
    get:
      summary: List staff time off (overlapping range)
      security:
        - bearerAuth: []
      parameters:
        - name: staff_id
          in: query
          required: true
          schema:
            type: string
        - name: from
          in: query
          required: true
          schema:
            type: string
            format: date-time
        - name: to
          in: query
          required: true
          schema:
            type: string
            format: date-time
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/TimeOff"
              examples:
                list:
                  value:
                    - id: "1f6b7a2c-9a3b-4a1e-8b2c-4c5d6e7f8a90"
                      staff_id: "2d7f53f6-0b5f-4d0d-8f49-7a9c3b3b9c2a"
                      start_time: "2026-02-01T11:00:00Z"
                      end_time: "2026-02-01T12:00:00Z"
                      reason: "Doctor appointment"
                      created_at: "2026-01-28T10:05:00Z"
    delete:
      summary: Delete staff time off (by id)
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: query
          required: true
          schema:
            type: string
      responses:
        "204":
          description: No Content
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    BusinessProfile:
      type: object
      properties:
        business_id:
          type: string
        name:
          type: string
        timezone:
          type: string
        reminder_offsets_minutes:
          type: array
          items:
            type: integer
    BusinessProfileUpdateRequest:
      type: object
      properties:
        name:
          type: string
        timezone:
          type: string
        reminder_offsets_minutes:
          type: array
          items:
            type: integer
    BusinessServiceCreateRequest:
      type: object
      required: [name, duration_minutes, price]
      properties:
        name:
          type: string
        duration_minutes:
          type: integer
        price:
          type: number
          format: float
        description:
          type: string
    BusinessService:
      type: object
      properties:
        id:
          type: string
        business_id:
          type: string
        name:
          type: string
        duration_minutes:
          type: integer
        price:
          type: string
        description:
          type: string
        created_at:
          type: string
          format: date-time
    StaffCreateRequest:
      type: object
      required: [name]
      properties:
        name:
          type: string
        is_active:
          type: boolean
    Staff:
      type: object
      properties:
        id:
          type: string
        business_id:
          type: string
        name:
          type: string
        is_active:
          type: boolean
    WorkingHours:
      type: object
      properties:
        staff_id:
          type: string
        weekday:
          type: integer
        is_working:
          type: boolean
        start_minute:
          type: integer
        end_minute:
          type: integer
    WorkingHoursUpsertRequest:
      type: object
      required: [weekday, is_working]
      properties:
        weekday:
          type: integer
        is_working:
          type: boolean
        start_minute:
          type: integer
        end_minute:
          type: integer
    IdResponse:
      type: object
      properties:
        id:
          type: string
    TimeOffCreateRequest:
      type: object
      required: [start_time, end_time]
      properties:
        start_time:
          type: string
          format: date-time
        end_time:
          type: string
          format: date-time
        reason:
          type: string
    TimeOff:
      type: object
      properties:
        id:
          type: string
        staff_id:
          type: string
        start_time:
          type: string
          format: date-time
        end_time:
          type: string
          format: date-time
        reason:
          type: string
        created_at:
          type: string
          format: date-time
...